move sendgrid client out of civil package
