+++ commons-civil/src/main/java/uk/gov/hmcts/reform/sendgrid/EmailSendFailedException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.sendgrid;
+package uk.gov.hmcts.reform.sendgrid;
 
 public class EmailSendFailedException extends RuntimeException {
