+++ commons-civil/src/main/java/uk/gov/hmcts/reform/sendgrid/config/SendGridAutoConfiguration.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.sendgrid.config;
+package uk.gov.hmcts.reform.sendgrid.config;
 
 import com.sendgrid.SendGrid;
 import com.sendgrid.SendGridAPI;
