+++ commons-civil/src/main/java/uk/gov/hmcts/reform/sendgrid/config/SendGridProperties.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.sendgrid.config;
+package uk.gov.hmcts.reform.sendgrid.config;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
