+++ commons-civil/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  uk.gov.hmcts.reform.civil.sendgrid.config.SendGridAutoConfiguration
+  uk.gov.hmcts.reform.sendgrid.config.SendGridAutoConfiguration
