+++ commons-civil/src/test/java/uk/gov/hmcts/reform/sendgrid/EmailDataTest.java
@@ -1,8 +1,10 @@
-package uk.gov.hmcts.reform.civil.sendgrid;
+package uk.gov.hmcts.reform.sendgrid;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import uk.gov.hmcts.reform.sendgrid.EmailAttachment;
+import uk.gov.hmcts.reform.sendgrid.EmailData;
 
 import java.util.List;
